Remove unused function parameter
The function getContributionParams was created as part of a larger refactoring/
code cleanup effort. One of those goals was to get to a point
where contributions are always created with a pending status and are later
confirmed when a payment is added.

It if only called from one place and TRUE is always passed for
so we no longer need the  parameter in this function, remove it