Fix & test searchQuery order by to be less dependent on what is selected for search
The pseudoconstant array is built when doing a select - if a field is in
order by but not select it will fail - as in the test.

This is a bit hard to hit outside the test for contribution but makes it more robust`
allowing us to address other bugs & performance issues.