Support calling ContributionPage.validate in POST context.
Because the validation data is buried deep within the form we need to load the form
in order to validate - but currently the validKey check borks in post requests.

Note the validity of credit card details is not checked when submitting this form,
other than possibly the luhn validation.

 A likely use case for the validation is when the is being submitted and some handling is
 to be done before processing but the validity of input needs to be checked first.

 For example Paypal Checkout will replace the confirm button with it's own but we are able to validate
before paypal launches it's modal. In this case the  is post but we need validation to succeed.