Fix relative dates when searching on activity_date_time
In testing I found that there is an unreleased regression relating to our date conversion
on activity_date_time in activity search. This fixes and adds tests. (regression was it
was being ignored).

Note that the option value is capitalised which makes the capitalisation slightly odd but
I think accepting that is the least bad option