CRM/Admin/Form/Setting/UF.tpl - Change "add" to "add or update"
If an admin follows the advice literally (and merely *adds* the item),
and if there's already an existing as directive higher-up (as
there would be in a stock Backdrop `settings.php`), then the
`settings.php` would have two copies of the directive.

That'd be liable to cause confusion during subsequent maintenance.