(NFC, infra/ops#878) AssetBuilderTest - Relax getUrl() assertion for WP compat
The test ensures that AssetBuilder produces plausible URLs when caching is disabled.

Generated URLs look different in the default `drupal-clean` and `wp-demo` configurations, e.g.

* D7 (drupal-clean): `http://dmaster.bknix:8001/civicrm/asset/builder?an=square.js&ap=...`
* WP (wp-demo): `http://wpmaster.bknix:8001/?page=CiviCRM&q=civicrm%2Fasset%2Fbuilder&an=square.js&ap=...

The assertion in `AssetBuilderTest` was coded in a way that fails on working URLs.

Before
------

The test fails on WP.

After
-----

The test passes on WP.