dev/core#755 - CRM/Core - Fix missing group name in custom field cache
CRM_Core_BAO_CustomField::getCustomFieldID caches custom fields based
on the field name, meaning if a custom field of the same name exists
in more than one group, the method will always return the first field
that was cached. This change adds the $groupName parameter to the
cache key.

This also adds a missing test for the $fullString parameter and
renames the $groupTitle parameter to $groupName in preparation
for future changes.