dev/financial#2 add PaymentProcessor.title field
There was a long discussion a while back about adding a label or title field to the
civicrm_payment_processor table. It kinda died but since 5.13 adds another field
(contribution_recur.cancel_reason) which is kinda rare I thought we should probably
add this field in the same release as we can better keep most releases to no
schema changes that way.

At this stage the field is not in use - but I figure getting it added will make the
next steps of exposing it easier & there is general agreement we need
something of this nature.

(the cancel reason field will also take further commits to expose in the UI)