[REF] extract token functions
This is an reviewer's commit from https://github.com/civicrm/civicrm-core/pull/12012/files

We are about to merge array formatting changes that will make lots of PRs go stale.

I couldn't get this reviewed & merged before the change but I thought if I could do it
through a sub-commit it would be better than just making it go stale.

This is a simple extraction & I will add merge on pass as it is a reviewer's commit