(REF) crmMailingAB - Use the new `winner_id` option instead of `crmMailingMgr.mergeInto`
In this context, the user has run an A/B experiment and chosen a specific
variant (the "winner") to use in the full blast.  The purpose of the code is
to copy the "winner" content to the final mailing - and send the final
mailing.

This replaces some client-side logic with a server-side equivalent.