Switch to using ContributionRecur.cancel api from CancelSubscription form
The function cancelRecurContribution is only called from the form ant from the api.

This makes it so the form uses the api, which we have started to add unit testing to. Follow
ing this we can use the api for transaction handling and move more of the form logic into
the api, making it available to others & allowing us to better move to a new form layer later
without as much logic in the forms