[REF] remove fields from input parameters of formatLocationBlock
As part of a tidy up of this function I am removing  as an input parameter & getting it from
a function. Unit testing added in testAddressWithCustomData to ensure custom data is still imported.

Additional related tests are in that function. This is preliminary to fixing
primary handling & related test for phone & email import - and to tidying up general mess in this function