Update dao.tpl to not make presence of import or export required to add 'where' param
Import & export have been hopelessly overloaded & misued in the code. One reason is that without them
'where' is not present. Logically import & export should have a narrow usage and should not
be a pre-requisite for being available as a search field, in reality not so much.

This simply makes the 'where' field available regardless of the presence of these fields

By doing this we can make the metadata in the Query object able to handle more fields
without special handling (e.g Contribution Recur fields)