Move the qfbug handling to it's own function
This is a really toxi function. In order to save the pain or iterating multiple times through a small array
we are dealing with an awful lot of complexity. I'm changing this to iterate through it multiple
times in order to rationalise the code - I think performance is minimal & being able to read the code....