(flexmailer#29) civicrm/mailing/view - Generate content via Mailing.preview API
A root cause of flexmailer#29 is that the flexmailer has to override
multiple parts of CiviMail.  Case in point: it overrides the
`civicrm/mailing/view` and forces it to generate content via
`Mailing.preview` API.  This is unfortunate because flexmailer's variant is
missing other features (regarding permissioning and contact IDs).

This revision makes it unnecessary for flexmailer to override
`civicrm/mailing/view`.