[WIP] remove use of discontinued phpunit/dbunit
The phpunit/dbunit package has been discontinued. I've removed the reference to it from the top level
extends and one call to it & the test I tried locally passes so it's not needed for ALL tests.

Let's see where it IS needed & what our options are - have at it jenkins