Reduce cache bypasses
We have been running this patch in production for some months. When you load up a contribution edit page
(for example) the custom data metadata is loaded without hitting the cache. This means that
unnecessary queries are done. The only reason to bypass the cache would be that it is stale
or parameter dependent. However, the staleness is/ should be dealt with when editing custom
fields & if it is not then there will be other impacts. The entity id is NOT passed into the part of
the code that goes into the cached value