Skip wrapping v4 api requests in a transaction
Transactions aren't a bad thing and we probably want to revisit this but
for now I'm disabling it becuase it's broken, as seen in api_v3_ActivityTest.
I think the problem is with the onApiException event not being dispatched
because api4 doesn't go through Kernel->runSafe, so transactions aren't
being rolled back when they should.