dev/core#980 update phpunit extends class to support later versions
Per https://engineering.facile.it/blog/eng/phpunit-upgrade-namespace/ this works with phpunit 4.8 +

Phpunit support for php versions is here https://phpunit.de/supported-versions.html

Basically we need phpunit 7 to support php 7.3 - which we should be adding to our
test suite now. It's hard to make a case that we still need to support phpunit 4.8 - let
alone versions before that - 5.7 still covers php 5.6 which we need for some of our current branches.

Unit testing on master would ideally switch up to a supported version of phpunit - 7 or 8