dev/core#905 Separate contribution_recur status_id option group from contribution option group
As we did earlier for pledge status this separates out the option groups - on the basis we have different
statuses for the 2 things. In order to ensure we don't break anything existing the values
we give the new option group the exact same values as the old one>

However, I HAVE added one more value - Processing - which is intended for the period between when
a new contribution starts to be added and when it has been added. This would normally be a few seconds
but when there is an error it would be left in this state - warding off future attempts.

For new sites I have left off some statuses that I think would not be relevant - Refunded,
Partially Paid, Chargeback, Pending Refund