Add default receive_date for contributions at BAO level
I found that CRM_Contribute_Page_AjaxTest actually fails if run in isolation - obviously 'something' is
helping it out when run together but in isolation the receive_date is missing on the contribution
which causes an error when it goes to save financial item. (There was a similar test issue recently
that turned out to impact form users - this solves the issue deeper down. The default is only applied
if id is not present