dev/core#1025 fix unreleased regression on contact.get with custom fields
Fixes a bug where contact.get does not retrieve results if a custom field is requested and an activity custom field exists

This bug came about because a recent fix made metadata more consistent but it turns out this class was filtering the
metadata by the passed in  - and then iterating through the metadata. Now it iterates the ids instead.