Copy assignProportional Line items back into Payment.create function
This is a case where (unintuitively) I feel reducing code sharing and extraction makes sense. The reason being
I think the 'needs' of adding a payment are not obviously the same as the needs of 'changing a payment instrument'
and the code feels confusion and like the extractions make less sense now than they did. In the interests
of improving it bringing it back together seems the right path to me here. We can then edit for the purposes
of payment create code very specifically (and all functions that add payments need to call that)