Remove cached smart group entries when removing from a group
In 5.15 we have added fixes for smart groups regarding removed entries. Possibly related to this but also possibly pre-existing
I'm seeing contacts who have been removed from smart groups showing up after removal in search results for the duration of the smart group cache.

Regardless the fix feels safe, sensible and correct so I'm comfortable targetting the rc without further research as to whether it
is a regression. The fix being removal of the cache entry when the group entry is removed.