Specifiy allowed_classes on unserialize
This is in response to phpstorm grumpily pointing to
https://github.com/kalessil/phpinspectionsea/blob/master/docs/security.md#exploiting-unserialize
in it's inspections. I think in this case it is very much cosmetic. However,
I wonder is we should have a function
CRM_Utils_String::unserialize() that does this by default & becomes the place we
add other handling if we identify any. (It might also make it easier to migrate to
json at some point