[REF] move formatting of conflict into markConflict function
The goal here is to make this information more usable to an api caller. Currently the 'conflict' key holds text formatted
for a specific form. I was going to change this to hold an array of data that an api could use. However, the
Contact.get_merge_conflicts already has a specific api-friendly format that is keyed by 'mode'
at the top level so my revised plan is to return api-friendly data as well (without messign with the
existing form's data).

This change simply moves the decisions about what data to store & how down to the function whose responsibility
it is (markConflict) and adds a test to ensure that no change results