Consolidate handling of conflicts between the batch job and get_conflicts api
This ensures that conflicts are stored during batch_merge to the prev_next cache with the same format as when the api
calls get_conflicts. The code doing this wrangling is moved from the api to the BAO layer.

We add a test to ensure the output is the same & use the previously added test to check the string is the same too.

Test cover here is pretty good