Fix failing test by changing expected date format
We have been seeing failures on testCancelledFilter on day one of the month - the reason seems to be
that we are expecting, but not receiving, a date format with a leading zero in the qill. This would
pass for much of the month but fail during the single digit days. This doesn't seem like a regreession to
me - just a test-error surfaced by changes in https://github.com/eileenmcnaughton/civicrm-core/commit/6e7cc0f578e814459b1e1b80c7a922684da46140#diff-984fd916f4708bb4b370841ae0ab69e5R465