e-notice fix & unit test
This addresses a fairly obscure enotice. The notice occurs when a table has been excluded from
triggers by a hook (as makes sense if the table includes only calculated fields) and then a new
custom field is added to the table (or presumably edited or removed). In this case we should
not go through the whole trigger handling loop.

Note I see it as a feature rather than a bug that this would also apply to triggers that
update the contact.modified_date as it's not helpful to update this when updating a
calculation - if someone decides to intervene by hook we should trust them