[REF] un-extract createProportionalFinancialEntities
This is one of those places where, unituitively, I think undoing an extraction makes sense.

The code for allocating monies to financial items is shared with the change payment instrument functionality
but fundamentally these are different processes and sharing the code is making it less rather than more readable.

My goal is to pull back all the code that handles processing financials off a payment back into Payment.create
and to clean it up from there