dev/core#1108 [REF] use CRM_Core_DAO::executeQuery instead of ->query()
This allows us to leave the multilingual translation to the I18n rewriter. This PR just removes 2 instances from this function.

Once the others are removed we should be able to remove the swath of code defining I18n tables are the start of the function
and simplify the tables involved

Note testUnsubscribeGroupList passes through this code