[REF] [export] remove chunk of non-functional code
I worked through this & groups is only available as a return property from contact export so can only be hit from there.

However, the search string has a very specific format - involving parentheses - that is NOT present in the from
clause as currently calculated

      case 'civicrm_group_contact':
        return " $side JOIN civicrm_group_contact ON contact_a.id = civicrm_group_contact.contact_id ";

I don't kmow if there is some bug that if this worked it would fix or not but the original bug
it addressed referred to an old query format we don't use https://issues.civicrm.org/jira/browse/CRM-3618 now.

If there is a bug this won't change it & if we identify it we can add tests & fix properly